requirements.txt
streamlit
langchain_huggingface
python-dotenv
pandas
plotly